// envs/figma/mock/src/types/ops.ts
export interface Viewport {
  x: number;
  y: number;
  zoom: number;
}